refactor: move interfaces.ts to the types folder
